Update bootstrap_moonlight-qt.sh
Not sure what this change does from a proper point. But commenting it out makes moonlight launch on my Linux Mint 20.2 with kodi 19.4

Thank you for this, i was in the early stages of removing the docker parts of this add on so i could run it on my box. I will help where and if you need me to.